Add files via upload
Test code to look left and right. I commented out the last part that centers the eyes since we don't need that for the test. Also I found out that when it looks right, it goes to 280. I'm assuming to position 280 degrees? And to look left it goes to 380 (degrees??). Which means we get 100 degrees for Field of View. So it can turn 50 degrees in either direction.